Improve DB init script
Put reasonable limits to the columns (not just 255 for all). Make insert
statements more explicit about which column a field of the data belongs
to. Decide not to make initial data for the submissions table since we need
the data for the file column and since it doesn't make sense to pre-populate the
submissions. Finally, decide not to use MySQL's DATETIME format since it
does not support ISO 8601 Datetime format---use VARCHAR instead; the
conversion from string to ISO 8601 will be done by JavaScript.