Users endpoints (#10)
* added changes phi made

* post user working

* need to add in stuff for instructor and student info, but going to wait for their corresponding models to be ready as not to duplicate work. Otherwise good to go!

* fixed auth

* added in student and instructor roles

* Make modifications before merge into master

- Most of the changes are to make the code compliant with workspace's
ESLint.
- A primary change is to clear up confusion between `user`, `req.user`,
and the user object for authenticated user, etc. by assigning different
User objects with better names. I.e. to have `currentUser` (the currently logged
in user), `requestedUser` (the user whose ID is in the route), and
`req.authenticatedUserId` (the ID of the successfully authenticated token).
- Change the logic of the GET /users/{id} a litte bit, since the old
version never actually reached 404 (due to it always trying to matching
the current logged in user's ID with the one in the request URL, and
since there's no way to log in into users that don't exist, it always
returned 403).
- Update the `UserSchema`. The `id` field is not required. On successful
database insertion, MySQL will automatically assign an ID for the newly
inserted row.
- Add the option to filter out user's password directly in the
data-access methods.
- If your function already returns a Promise, I don't think you need to specify
it to be an async function.
- If you want to use `req.params.id` as an integer multiple times, you can just
parse it in the beginning of the middleware, e.g.
`const id = parseInt(req.params.id);` and then use `id` throughout the function.